+++ src/out/ci/build.sh
@@ -1,3 +1,3 @@
 cd /home/user/lnx-build
-sudo chmod 775 cs-script_4.8-6/DEBIAN/p*
-dpkg-deb --build cs-script_4.8-6
+sudo chmod 775 cs-script_4.8-7/DEBIAN/p*
+dpkg-deb --build cs-script_4.8-7